move non-user traits class files into internal
