+++ Periodic_3_triangulation_3/include/CGAL/internal/Periodic_3_Delaunay_triangulation_statically_filtered_traits_3.h
@@ -30,7 +30,7 @@
 // logic and is not plug'n play (requires users providing bounds).
 // If it should be provided again, it should probably be separate.
 
-#include <CGAL/Periodic_3_triangulation_statically_filtered_traits_3.h>
+#include <CGAL/internal/Periodic_3_triangulation_statically_filtered_traits_3.h>
 
 #include <CGAL/internal/Static_filters/Periodic_3_side_of_oriented_sphere_3.h>
